requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
numpy
datasets
boto3